disable cancel button for non active tasks; move makeSubmitTask to test-data
