+++ packages/react-components/tests/test-data/tasks.ts
@@ -1,4 +1,4 @@
-import { TaskProgress } from 'api-client';
+import type { SubmitTask, TaskProgress } from 'api-client';
 import * as RmfModels from 'rmf-models';
 
 export function makeTask(
@@ -24,6 +24,17 @@ export function makeTask(
     fleet_name: 'test_fleet',
     robot_name: 'test_robot',
   });
+}
+
+export function makeSubmitTask(): SubmitTask {
+  return {
+    description: {
+      cleaning_zone: 'zone',
+    },
+    start_time: Math.floor(Date.now() / 1000),
+    task_type: RmfModels.TaskType.TYPE_CLEAN,
+    priority: 0,
+  };
 }
 
 export function makeDefinedTask(
